+++ web-app/src/main/java/com/lyna/web/domain/logicstics/repository/LogiticDetailRepositoryImpl.java
@@ -20,14 +20,12 @@ public class LogiticDetailRepositoryImpl extends BaseRepository<LogiticsDetail, 
     }
 
     @Override
-    public boolean deleteByPackageIds(List<String> packageIds) throws DomainException {
-        try {
-            String query = "DELETE FROM LogiticsDetail l WHERE l.packageId in (:packageIds)";
-            entityManager.createQuery(query).setParameter("packageIds", packageIds).executeUpdate();
+    public boolean deleteByPackageIdsAndTenantId(List<String> packageIds, int tenantId) throws DomainException {
+            String query = "DELETE FROM LogiticsDetail l WHERE l.packageId in (:packageIds) AND l.tenantId=:tenantId";
+            entityManager.createQuery(query)
+                    .setParameter("packageIds", packageIds)
+                    .setParameter("tenantId", tenantId)
+                    .executeUpdate();
             return true;
-        } catch (IllegalArgumentException e) {
-            log.error(e.getMessage());
-            return false;
-        }
     }
 }
