+++ web-app/src/main/java/com/lyna/web/domain/mpackage/repository/impl/PackageRepositoryImpl.java
@@ -4,6 +4,7 @@ import com.lyna.commons.infrustructure.exception.DomainException;
 import com.lyna.commons.infrustructure.repository.BaseRepository;
 import com.lyna.web.domain.mpackage.Package;
 import com.lyna.web.domain.mpackage.repository.PackageRepository;
+import com.lyna.web.domain.stores.exception.StoreException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
@@ -23,33 +24,28 @@ public class PackageRepositoryImpl extends BaseRepository<Package, String> imple
 
     @Override
     @Transactional
-    public Package findOneByPackageId(String packageId) throws DomainException {
-        try {
+    public Package findOneByPackageIdAndTenantId(String packageId, int tenantId) throws DomainException {
             return entityManager
-                    .createQuery("SELECT p FROM Package p WHERE p.packageId=:packageId", Package.class)
+                    .createQuery("SELECT p FROM Package p WHERE p.packageId=:packageId AND p.tenantId=:tenantId", Package.class)
                     .setParameter("packageId", packageId)
+                    .setParameter("tenantId", tenantId)
                     .getSingleResult();
-        } catch (Exception e) {
-            log.error(e.getMessage());
-            return null;
-        }
     }
 
     @Override
-    public boolean deleteByPackageIds(List<String> packageIds) throws DomainException {
-        try {
-            String query = "DELETE FROM Package p WHERE p.packageId in (:packageIds)";
-            entityManager.createQuery(query).setParameter("packageIds", packageIds).executeUpdate();
+    public boolean deleteByPackageIdsAndTenantId(List<String> packageIds, int tenantId) throws DomainException {
+            String query = "DELETE FROM Package p WHERE p.packageId in (:packageIds) AND p.tenantId=:tenantId";
+            entityManager.createQuery(query)
+                    .setParameter("packageIds", packageIds)
+                    .setParameter("tenantId", tenantId)
+                    .executeUpdate();
             return true;
-        } catch (Exception e) {
-            log.error(e.getMessage());
-            return false;
-        }
     }
 
     @Override
     @Transactional
     public List<Package> findByTenantId(int tenantId) {
+
         return entityManager
                 .createQuery("SELECT p FROM Package p WHERE p.tenantId=:tenantId order by p.name", Package.class)
                 .setParameter("tenantId", tenantId)
